requirements-photo-migration.txt
# Dependências Python para o script de migração de fotos
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.4
